Move BDN9 rev 1 to subfolder
