+++ keyboards/keebio/bdn9/bdn9.h
@@ -1,37 +1,5 @@
-/* Copyright 2019 Danny Nguyen <danny@keeb.io>
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 2 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
 #pragma once
 
-#include "quantum.h"
-
-/* This a shortcut to help you visually see your layout.
- *
- * The first section contains all of the arguments representing the physical
- * layout of the board and position of the keys.
- *
- * The second converts the arguments into a two-dimensional array which
- * represents the switch matrix.
- */
-#define LAYOUT( \
-    KA1, KA2, KA3, \
-    KB1, KB2, KB3, \
-    KC1, KC2, KC3 \
-) \
-{ \
-    { KA1, KA2, KA3 }, \
-    { KB1, KB2, KB3 }, \
-    { KC1, KC2, KC3 } \
-}
+#ifdef KEYBOARD_keebio_bdn9_rev1
+    #include "rev1.h"
+#endif
